pyproject.toml
[tool.poetry]
name = "amyrin"
version = "0.1.0"
description = "cool discord bot"
authors = ["syrice"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.2.2"
pyyaml = "^6.0"
jishaku = "^2.5.1"
gitpython = "^3.1.31"
humanfriendly = "^10.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
